Converting to typescript
