+++ react/components/Menu/MenuLinksList.tsx
@@ -8,7 +8,7 @@ import {
   defineMessages,
   InjectedIntlProps,
 } from 'react-intl'
-import { ExtensionPoint } from 'render'
+import { ExtensionPoint } from 'vtex.render-runtime'
 import { AuthService } from 'vtex.react-vtexid'
 import { ModalDialog } from 'vtex.styleguide'
