+++ react/index.tsx
@@ -6,8 +6,6 @@ import { logMyAccountURL, logGeneralErrors } from './utils/splunk'
 
 import 'vtex.country-codes/locales'
 
-import styles from './styles.css'
-
 class MyAccount extends Component {
   public constructor(props: any) {
     super(props)
