+++ react/typings/css.d.ts
@@ -1,0 +1,7 @@
+declare module '*.css' {
+  interface ClassNames {
+    [className: string]: string
+  }
+  const classNames: ClassNames
+  export = classNames
+}
